Autenticação e Autorização com JWT

+++ tests/test_app.py
@@ -76,15 +76,10 @@ def test_read_user(client, user):
     assert response.json() == user_schema
 
 
-def test_read_user_404(client):
-    response = client.get('/users/1024')
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'User not found'}
-
-
-def test_update_user(client, user):
+def test_update_user(client, user, token):
     response = client.put(
         f'/users/{user.id}',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'bob',
             'email': 'bob@example.com',
@@ -92,10 +87,10 @@ def test_update_user(client, user):
         },
     )
     assert response.status_code == HTTPStatus.OK
-    assert response.json() == {'username': 'bob', 'email': 'bob@example.com', 'id': 1}
+    assert response.json() == {'username': 'bob', 'email': 'bob@example.com', 'id': user.id}
 
 
-def test_update_user_integrity_error(client, user):
+def test_update_user_integrity_error(client, user, token):
     client.post(
         '/users',
         json={
@@ -107,6 +102,7 @@ def test_update_user_integrity_error(client, user):
 
     response_update = client.put(
         f'/users/{user.id}',
+        headers={'Authorization': f'Bearer {token}'},
         json={
             'username': 'pedro',
             'email': 'mateus@gmail.com',
@@ -118,26 +114,16 @@ def test_update_user_integrity_error(client, user):
     assert response_update.json() == {'detail': 'Username or email already exists'}
 
 
-def test_update_user_404(client):
-    response = client.put(
-        '/users/1024',
-        json={
-            'username': 'bob',
-            'email': 'bob@example.com',
-            'password': 'mynewpassword',
-        },
-    )
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'User not found'}
-
-
-def test_delete_user(client, user):
-    response = client.delete(f'/users/{user.id}')
+def test_delete_user(client, user, token):
+    response = client.delete(f'/users/{user.id}', headers={'Authorization': f'Bearer {token}'})
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {'message': 'User deleted'}
 
 
-def test_delete_user_404(client):
-    response = client.delete('/users/1024')
-    assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == {'detail': 'User not found'}
+def test_get_token(client, user):
+    response = client.post('/token', data={'username': user.username, 'password': user.clean_password})
+    token = response.json()
+
+    assert response.status_code == HTTPStatus.OK
+    assert 'access_token' in token
+    assert 'token_type' in token
